+++ src/os2datascanner/projects/admin/organizations/views/account_views.py
@@ -9,7 +9,7 @@ from django.forms import ModelForm
 from django.shortcuts import get_object_or_404
 from django.utils.translation import gettext_lazy as _
 from django.db.models.functions import Concat, Greatest
-from django.db.models import CharField, Value
+from django.db.models import CharField, Value, Max
 
 from ..models import Account, Alias, OrganizationalUnit
 from ..models.aliases import AliasType
@@ -58,7 +58,15 @@ class AccountListView(ClientAdminMixin, RestrictedListView):
                     TrigramSimilarity("full_name", search),
                     TrigramSimilarity("username", search)
                 )
-            ).filter(search__gte=0.2)
+            ).exclude(search=0)
+            if qs.exists():
+                max_similarity = qs.aggregate(Max("search", default=0))['search__max']
+                # We only want accounts whose similarity to the search word is at least,
+                # 0.3 times as similar as the most similar one.
+                # As such a user won't recieve an empty list of results.
+                # 0.3 has been chosen, as this threshold should include all instances of a name,
+                # including different spellings, such as "Christopher", "Cristofer", etc.
+                qs = qs.filter(search__gte=max_similarity * 0.3)
 
         return qs